Move options mapper to flags.go
